Followed naming conventions, added comments for components, used variables in stylesheets
