+++ package.json
@@ -9,6 +9,8 @@
   "dependencies": {
     "react": "15.4.2",
     "react-native": "0.42.3",
+    "react-native-calendar": "^0.11.0",
+    "react-native-checkbox": "^1.1.0",
     "react-native-modal-dropdown": "^0.4.2",
     "react-native-selectme": "^1.2.3",
     "react-native-touchable-set-active": "0.0.5"
