--- src/components/Calender/style.js
@@ -1,31 +1,0 @@
-import { StyleSheet, Dimensions } from 'react-native';
-
-export default StyleSheet.create({
-    calendarContainer: {
-        backgroundColor: '#303b3d',
-    },
-    currentDayCircle: {
-        backgroundColor: '#ff9f02',
-    },
-    day: {
-        color: '#fff',
-    },
-    currentDayText: {
-        color: '#303b3d',
-    },
-    title: {
-        color: '#fff',
-    },
-    controlButtonText: {
-        color: '#fff',
-    },
-    calendarHeading: {
-        borderTopColor: '#ff9f02',
-        borderWidth: 1,
-    },
-    weekRow: {
-        borderTopColor: '#ff9f02',
-        borderBottomColor: '#ff9f02',
-        borderWidth: 1,
-    },
-});
