--- src/components/Container/index.js
@@ -1,18 +1,0 @@
-import React, { Component } from 'react';
- 
-import {
-  StyleSheet,
-  View
-} from 'react-native';
-import styles from './style';
- 
-export default class Container extends Component {
-
-    render(){
-        return (
-            <View style={styles.labelContainer}>
-                { this.props.children }
-            </View>
-        );
-    } 
-}
