--- src/components/Container/style.js
@@ -1,7 +1,0 @@
-import { StyleSheet, Dimensions } from 'react-native';
-
-export default StyleSheet.create({
-    labelContainer: {
-        marginBottom: 20
-    }
-});
