--- src/components/Footer/style.js
@@ -1,15 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-    footerContainer: {
-  		flexDirection: 'row',
-		justifyContent: 'space-between',
-        alignItems: 'center',
-        height: 40,
-        padding: 10,
-        backgroundColor: '#293334',
-        borderBottomColor: '#46575a',
-        borderStyle: 'solid',
-        borderWidth: 1,
-    },
-});
