--- src/components/Ticker/style.js
@@ -1,10 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-    tickImg: {
-        width: 15,
-        height: 15,
-        marginRight: 10,
-        padding:5,
-    },
-});
