--- src/components/TouchableHighlightBtn/style.js
@@ -1,9 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-    button: {
-        alignItems: 'center',
-        justifyContent: 'center',
-        height: 50,
-    },
-});
