--- src/components/TouchableOpacityBtn/style.js
@@ -1,7 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-	labelStyle: {
-		color: '#ff9f02'
-	}
-});
