+++ src/components/highlight-btn/style.js
@@ -1,0 +1,12 @@
+/**
+ * Hightlight button styles
+**/
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+    button: {
+        alignItems: 'center',
+        justifyContent: 'center',
+        height: 50,
+    },
+});
