+++ src/components/opacity-btn/style.js
@@ -1,0 +1,12 @@
+
+/**
+ * Hightlight button styles
+**/
+var labelColor    =  '#ff9f02';
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	labelStyle: {
+		color: labelColor
+	}
+});
