+++ src/components/ticker/style.js
@@ -1,0 +1,13 @@
+/**
+ * RememberMe styles
+ **/
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+    tickImg: {
+        width: 15,
+        height: 15,
+        marginRight: 10,
+        padding:5,
+    },
+});
