+++ src/components/wrapper/style.js
@@ -1,0 +1,10 @@
+/**
+ * Wrapper styles
+ **/
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+    labelContainer: {
+        marginBottom: 20
+    }
+});
