+++ src/views/MainCategoryView/style.js
