+++ src/views/SetInfoView/style.js
